Slight bugfixes
