--- Assets/Develop/Scripts/GroundDetectorV2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9e9c8fa0fa948f040bc280b349b7ea47
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
