+++ Assets/Develop/Scripts/WallDetector.cs
@@ -2,30 +2,34 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-public class WallDetector : MonoBehaviour {
+public class WallDetector : MonoBehaviour
+{
 
-    private MovementV2 playerMovement;
+    private Movement playerMovement;
 
-	// Use this for initialization
-	void Start () {
-        playerMovement = transform.parent.GetComponent<MovementV2>();
+    // Use this for initialization
+    void Start()
+    {
+        playerMovement = transform.parent.GetComponent<Movement>();
     }
 
-    private void OnTriggerEnter2D(Collider2D col) {
-        playerMovement.GetComponent<MovementV2>().wallHit = true;
+    private void OnTriggerEnter2D(Collider2D col)
+    {
+        playerMovement.GetComponent<Movement>().wallHit = true;
 
         if (this.gameObject.name == "Wall Trigger Left")
         {
-            playerMovement.GetComponent<MovementV2>().wallTrigger = 1;
+            playerMovement.GetComponent<Movement>().wallTrigger = 1;
         }
         else
         {
-            playerMovement.GetComponent<MovementV2>().wallTrigger = -1;
+            playerMovement.GetComponent<Movement>().wallTrigger = -1;
         }
     }
 
-    private void OnTriggerExit2D(Collider2D col) {
-        playerMovement.GetComponent<MovementV2>().wallHit = false;
+    private void OnTriggerExit2D(Collider2D col)
+    {
+        playerMovement.GetComponent<Movement>().wallHit = false;
     }
 
 }
