--- Assets/Personals/Sondre/WallCollider.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3c0453ca2c306f44fad3dda5145b18ec
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
