--- Assets/Personals/Sondre/WallDetector.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 14a07d14067a30f4f8e21b3d318e4b66
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
